+++ wp-content/themes/whs-frame/footer.php
@@ -1,4 +1,8 @@
-<?php do_action( 'whs_frame_before_footer' ); ?>
+<?php
+defined( 'ABSPATH' ) || exit;
+
+do_action( 'whs_frame_before_footer' );
+?>
 <?php do_action( 'whs_frame_footer' ); ?>
 <?php do_action( 'whs_frame_after_footer' ); ?>